requirements.txt
backcall==0.2.0
certifi==2020.6.20
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.2.0
loguru==0.5.3
louvain==0.6.2
matplotlib==3.3.2
networkx==2.5
netwulf==0.1.5
numpy==1.19.2
pandas==1.1.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2020.1
scipy==1.5.2
seaborn==0.11.0
simplejson==3.17.2
six==1.15.0
texttable==1.6.3
tqdm==4.50.0
traitlets==5.0.4
wcwidth==0.2.5
win32-setctime==1.0.2
